# swagger.yaml
openapi: 3.0.1
info:
  title: Web IoT Makers Challenge
  description: メイカーズチャレンジのアプリケーションサーバーのエンドポイント
  version: "1.0.0"
servers:
  - url: http://localhost:3000/api
    description: Local server
paths:
  /devices:
    get:
      summary: デバイス情報の取得
      security:
        - bearerAuth: []
      description: ユーザーが持つすべてのデバイスの温度、湿度、開閉の情報をユーザーIDで取得する
      parameters:
        - name: "userId"
          in: "query"
          required: true
          schema:
            type: integer

      responses:
        200:
          description: デバイス情報のリスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Device:
      type: object
      properties:
        deviceId:
          type: integer
          example: 1
        devicename:
          type: string
          example: "umbrella A"
        temperature:
          type: number
          format: float
          example: 24.5
        humidity:
          type: number
          format: float
          example: 60.2
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        isOpen:
          type: boolean
          example: false
        userId:
          type: integer
          example: 1
...